backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
python-dotenv
python-jose[cryptography]
